Changed base-board to JSON. Added support for Chicago.
The built-in board elements are now defined in JSON. The structure it
provides feels more appropriate for files that A) won't change often
(if ever), and B) users won't be modifying.

Also, Chicago is now supported! Its unique cofiguration (stations on a
specific track) meant that it took some special consideration. The core
change (which impacts all the code) is paths are retrieved via method
call, rather than accessing the object directly. This allows it to handl
Chicago paths differently. Accordingly, the Chicago tiles have their own
object.