Corrected handling of Chicago route.
When asking Chicago for its paths, we need to consider if we're starting
from the tile. If so, only report the paths extending from the station
on the tile. Additionally, when filtering paths based on containing a
Chicago station, be sure to check the exit cell for presence in the
path, too.