Filtering invalid routes.
The route finding does not take into account that Chicago's stations are
completely separate, unlike any other tile. As such, it requires some
filter to strip out paths which turn out to be invalid.

This check looks to see if Chicago is the only station on the route, and
if so, asks if the path from it to the station's neighbor overlaps with
the route. If not, then the route is invalid, and is discarded.