Fix a couple route issues.
Both involved Chicago, although one only sort of, and the other required
a few conditions to be met.

The first was the failure to load Chicago stations. This occurred because
any whitespace was not appropriately stripped before checking for Chicago
in a railroad's stations. This looks to have only imacted command-line use
of the library.

In the second, if a railroad had its Chicago station at D8, and there was
a tile at D8, which connected to C7 but not D8, then the route could be
wrong; it could pass through C7, regardless of whether there was a station
there. This boils down to an shortcoming in the Route object; you can't
easily figure out which cell comes after a given cell. So evaluation of
the route to just checked D8 was included in the route, not that it
immediately followed D6.

The fix worked around that by finding the subroute of length 3 beginning at
C5, and running comparisons against it.