Add support for Big 4 and Michigan Southern.
When no owner is specified, a station is placed in their home city; when
an owner is specified, gives that railroad a station in the home city.
If the railroad already has one there, assumes the user did this manually,
and skips adding the station.

If it's phase 4 and no owner is listed, then no token is placed, as
private company tokens are removed during phase 4.